Move Lesson.jsx to pages/lessons subdirectory
